+++ DeriSock/Model/SubAccountPortfolio.cs
@@ -5,27 +5,27 @@
   public class SubAccountPortfolio
   {
     [JsonProperty("available_funds")]
-    public double AvailableFunds { get; set; }
+    public decimal AvailableFunds { get; set; }
 
     [JsonProperty("available_withdrawal_funds")]
-    public double AvailableWithdrawalFunds { get; set; }
+    public decimal AvailableWithdrawalFunds { get; set; }
 
     [JsonProperty("balance")]
-    public double Balance { get; set; }
+    public decimal Balance { get; set; }
 
     [JsonProperty("currency")]
     public string Currency { get; set; }
 
     [JsonProperty("equity")]
-    public double Equity { get; set; }
+    public decimal Equity { get; set; }
 
     [JsonProperty("initial_margin")]
-    public double InitialMargin { get; set; }
+    public decimal InitialMargin { get; set; }
 
     [JsonProperty("maintenance_margin")]
-    public double MaintenanceMargin { get; set; }
+    public decimal MaintenanceMargin { get; set; }
 
     [JsonProperty("margin_balance")]
-    public double MarginBalance { get; set; }
+    public decimal MarginBalance { get; set; }
   }
 }
