+++ DeriSock/Model/UserPosition.cs
@@ -8,19 +8,19 @@
     ///   Average price of trades that built this position
     /// </summary>
     [JsonProperty("average_price")]
-    public double AveragePrice { get; set; }
+    public decimal AveragePrice { get; set; }
 
     /// <summary>
     ///   Only for options, average price in USD
     /// </summary>
     [JsonProperty("average_price_usd")]
-    public double AveragePriceUsd { get; set; }
+    public decimal AveragePriceUsd { get; set; }
 
     /// <summary>
     ///   Delta parameter
     /// </summary>
     [JsonProperty("delta")]
-    public double Delta { get; set; }
+    public decimal Delta { get; set; }
 
     /// <summary>
     ///   Direction: <c>buy</c>, <c>sell</c> or <c>zero</c>
@@ -32,37 +32,37 @@
     ///   Only for futures, estimated liquidation price
     /// </summary>
     [JsonProperty("estimated_liquidation_price")]
-    public double EstimatedLiquidationPrice { get; set; }
+    public decimal EstimatedLiquidationPrice { get; set; }
 
     /// <summary>
     ///   Floating profit or loss
     /// </summary>
     [JsonProperty("floating_profit_loss")]
-    public double FloatingProfitLoss { get; set; }
+    public decimal FloatingProfitLoss { get; set; }
 
     /// <summary>
     ///   Only for options, floating profit or loss in USD
     /// </summary>
     [JsonProperty("floating_profit_loss_usd")]
-    public double FloatingProfitLossUsd { get; set; }
+    public decimal FloatingProfitLossUsd { get; set; }
 
     /// <summary>
     ///   Only for options, Gamma parameter
     /// </summary>
     [JsonProperty("gamma")]
-    public double Gamma { get; set; }
+    public decimal Gamma { get; set; }
 
     /// <summary>
     ///   Current index price
     /// </summary>
     [JsonProperty("index_price")]
-    public double IndexPrice { get; set; }
+    public decimal IndexPrice { get; set; }
 
     /// <summary>
     ///   Initial margin
     /// </summary>
     [JsonProperty("initial_margin")]
-    public double InitialMargin { get; set; }
+    public decimal InitialMargin { get; set; }
 
     /// <summary>
     ///   Unique instrument identifier
@@ -86,67 +86,67 @@
     ///   Maintenance margin
     /// </summary>
     [JsonProperty("maintenance_margin")]
-    public double MaintenanceMargin { get; set; }
+    public decimal MaintenanceMargin { get; set; }
 
     /// <summary>
     ///   Current mark price for position's instrument
     /// </summary>
     [JsonProperty("mark_price")]
-    public double MarkPrice { get; set; }
+    public decimal MarkPrice { get; set; }
 
     /// <summary>
     ///   Open orders margin
     /// </summary>
     [JsonProperty("open_orders_margin")]
-    public double OpenOrdersMargin { get; set; }
+    public decimal OpenOrdersMargin { get; set; }
 
     /// <summary>
     ///   Realized funding in current session included in session realized profit or loss, only for positions of perpetual
     ///   instruments
     /// </summary>
     [JsonProperty("realized_funding")]
-    public double RealizedFunding { get; set; }
+    public decimal RealizedFunding { get; set; }
 
     /// <summary>
     ///   Realized profit or loss
     /// </summary>
     [JsonProperty("realized_profit_loss")]
-    public double RealizedProfitLoss { get; set; }
+    public decimal RealizedProfitLoss { get; set; }
 
     /// <summary>
     ///   Last settlement price for position's instrument, 0 if instrument wasn't settled yet
     /// </summary>
     [JsonProperty("settlement_price")]
-    public double SettlementPrice { get; set; }
+    public decimal SettlementPrice { get; set; }
 
     /// <summary>
     ///   Position size for futures size in quote currency (e.g. USD), for options size is in base currency (e.g. BTC)
     /// </summary>
     [JsonProperty("size")]
-    public double Size { get; set; }
+    public decimal Size { get; set; }
 
     /// <summary>
     ///   Only for futures, position size in base currency
     /// </summary>
     [JsonProperty("size_currency")]
-    public double SizeCurrency { get; set; }
+    public decimal SizeCurrency { get; set; }
 
     /// <summary>
     ///   Only for options, Theta parameter
     /// </summary>
     [JsonProperty("theta")]
-    public double Theta { get; set; }
+    public decimal Theta { get; set; }
 
     /// <summary>
     ///   Profit or loss from position
     /// </summary>
     [JsonProperty("total_profit_loss")]
-    public double TotalProfitLoss { get; set; }
+    public decimal TotalProfitLoss { get; set; }
 
     /// <summary>
     ///   Only for options, Vega parameter
     /// </summary>
     [JsonProperty("vega")]
-    public double Vega { get; set; }
+    public decimal Vega { get; set; }
   }
 }
